first commit
The java application retrieves and prints the current time, and it
computes a hash value using an evenly distributed binary tree and the
current time.

The binary is used to compute the hash value. Each of the internal
nodes map to a math operation. Each of the leaf nodes map to a value.
Both the math operation and the value depends on the index of the node
and the bits in the current time.